+++ src/constants.js
@@ -1,6 +1,8 @@
 export const ACCESS_TOKEN_KEY = 'crowdin-access-token';
 export const ORGANIZATION = 'crowdin-organization';
 export const PROJECT_ID = 'crowdin-project-id';
+
 export const SYMBOL_TYPE = 'symbol-override';
 export const TEXT_TYPE = 'text';
+
 export const PLUGIN_VERSION = '1.2.1';
