+++ src/util/file.js
@@ -1,0 +1,10 @@
+
+function getFileName(element) {
+    return `Sketch_${element.id}.html`;
+}
+
+function getDirectoryName(page) {
+    return `Sketch_${page.id}`;
+}
+
+export { getFileName, getDirectoryName };
