+++ src/util/html.js
@@ -36,60 +36,7 @@ function convertArtboardToHtml(page, artboard) {
         x: artboard.frame.width,
         y: artboard.frame.height
     };
-    const textElements = dom.find('Text', artboard)
-        .map(e => {
-            const textId = e.id;
-            const text = e.text;
-            let parent = e.parent;
-            let parentId = parent.id;
-            let x = e.frame.x;
-            let y = e.frame.y;
-            while (parentId !== artboard.id) {
-                x += parent.frame.x;
-                y += parent.frame.y;
-                parent = parent.parent;
-                if (!parent) {
-                    return null;
-                }
-                parentId = parent.id;
-            }
-            return { x, y, textId, text, e, type: TEXT_TYPE };
-        });
-    const artboardSymbols = dom.find('SymbolInstance', artboard);
-    const textsFromSymbols = artboardSymbols
-        .map(symbol => {
-            return symbol.overrides
-                .filter(override => override.affectedLayer.type === 'Text' && override.property === 'stringValue')
-                .map(override => {
-                    const text = override.value;
-                    const e = override.affectedLayer;
-                    let parent = symbol;
-                    let parentId = parent.id;
-                    let x = e.frame.x;
-                    let y = e.frame.y;
-                    const parentSymbols = [];
-                    parentSymbols.push(symbol.id);
-                    while (parentId !== artboard.id) {
-                        x += parent.frame.x;
-                        y += parent.frame.y;
-                        let current = parent;
-                        parent = parent.parent;
-                        if (parent.type === 'SymbolMaster') {
-                            parent = getSymbolIstance(artboardSymbols, current.id, parent.id);
-                            if (parent !== null) {
-                                parentSymbols.push(parent.id);
-                            }
-                        }
-                        if (!parent) {
-                            return null;
-                        }
-                        parentId = parent.id;
-                    }
-                    return { x, y, textId: parentSymbols.reverse().join('/') + '/' + override.id, text, e, type: SYMBOL_TYPE };
-                });
-        })
-        .reduce((x, y) => x.concat(y), []);
-    const allTexts = textElements.concat(textsFromSymbols).filter(el => el !== null);
+    const allTexts = getTextElementsInArtboard(artboard);
     let textHtml = '';
     allTexts.forEach(t => {
         let style = `position: absolute;top:${t.y}px;left:${t.x}px;`;
@@ -151,6 +98,63 @@ function convertArtboardToHtml(page, artboard) {
     return html;
 }
 
+function getTextElementsInArtboard(artboard) {
+    const textElements = dom.find('Text', artboard)
+        .map(e => {
+            const textId = e.id;
+            const text = e.text;
+            let parent = e.parent;
+            let parentId = parent.id;
+            let x = e.frame.x;
+            let y = e.frame.y;
+            while (parentId !== artboard.id) {
+                x += parent.frame.x;
+                y += parent.frame.y;
+                parent = parent.parent;
+                if (!parent) {
+                    return null;
+                }
+                parentId = parent.id;
+            }
+            return { x, y, textId, text, e, type: TEXT_TYPE };
+        });
+    const artboardSymbols = dom.find('SymbolInstance', artboard);
+    const textsFromSymbols = artboardSymbols
+        .map(symbol => {
+            return symbol.overrides
+                .filter(override => override.affectedLayer.type === 'Text' && override.property === 'stringValue')
+                .map(override => {
+                    const text = override.value;
+                    const e = override.affectedLayer;
+                    let parent = symbol;
+                    let parentId = parent.id;
+                    let x = e.frame.x;
+                    let y = e.frame.y;
+                    const parentSymbols = [];
+                    parentSymbols.push(symbol.id);
+                    while (parentId !== artboard.id) {
+                        x += parent.frame.x;
+                        y += parent.frame.y;
+                        let current = parent;
+                        parent = parent.parent;
+                        if (parent.type === 'SymbolMaster') {
+                            parent = getSymbolIstance(artboardSymbols, current.id, parent.id);
+                            if (parent !== null) {
+                                parentSymbols.push(parent.id);
+                            }
+                        }
+                        if (!parent) {
+                            return null;
+                        }
+                        parentId = parent.id;
+                    }
+                    return { x, y, textId: parentSymbols.reverse().join('/') + '/' + override.id, text, e, type: SYMBOL_TYPE };
+                });
+        })
+        .reduce((x, y) => x.concat(y), []);
+    return textElements.concat(textsFromSymbols).filter(el => el !== null);
+}
+
 //for nested symbols Sketch API returns SymbolMaster instead of SymbolInstance as a parent group
 //and therefore we cannot calculate (x,y) of the text
 //this function is a workaround to find proper parent element
@@ -185,4 +189,4 @@ function parseHtmlForText(html) {
     return result;
 }
 
-export { convertArtboardToHtml, convertOutsideTextToHtml, parseHtmlForText }
+export { convertArtboardToHtml, convertOutsideTextToHtml, parseHtmlForText, getTextElementsInArtboard };
